run wallet on windows

+++ clib/xDagWallet/src/client/utils/log.c
@@ -16,7 +16,7 @@
 #include "../common.h"
 
 #if defined(_WIN32) || defined(_WIN64)
-#include < time.h >
+#include <time.h >
 #include "../../win/wintime.h"
 #else
 #include <sys/time.h>
@@ -50,7 +50,7 @@ int xdag_log(int level, int err, char* file, int line, const char *format, ...)
 
 	int pos = 0;
 	char buffer[4096] = {0};
-	pos += sprintf(buffer, "[%012llx][%.4s][%s] %s:%d  ", (long long)(pthread_self()), lvl + 4 * level, tbuf, xdag_filename(file), line);
+	pos += sprintf(buffer, "[%012llx][%.4s][%s] %s:%d  ", pthread_self(), lvl + 4 * level, tbuf, xdag_filename(file), line);
 	
 	va_start(arg, format);
 	pos = vsprintf(buffer + pos, format, arg);
